[MIG] l10n_fr_department_delivery from v10 to v14
