+++ l10n_fr_department_delivery/__init__.py
@@ -1,3 +1,1 @@
-# -*- encoding: utf-8 -*-
-
 from . import models
