+++ l10n_fr_department_delivery/__manifest__.py
@@ -1,17 +1,17 @@
-# -*- coding: utf-8 -*-
-#    Copyright (C) 2015 Akretion (http://www.akretion.com)
-#    @author Alexis de Lattre <alexis.delattre@akretion.com>
+# Copyright 2015-2023 Akretion France (http://www.akretion.com)
+# @author Alexis de Lattre <alexis.delattre@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 {
     'name': 'French Departments - Delivery',
     'summary': 'Use French Departments in delivery costs',
-    'version': '10.0.0.1.0',
+    'version': '14.0.1.0.0',
     'category': 'French Localization',
     'author': "Akretion,Odoo Community Association (OCA)",
-    'website': 'http://www.akretion.com',
+    'maintainers': ['alexis-via'],
+    'website': 'https://github.com/OCA/l10n-france',
     'license': 'AGPL-3',
     'depends': ['l10n_fr_department', 'delivery'],
-    'data': ['views/delivery.xml'],
+    'data': ['views/delivery_carrier.xml'],
     'installable': True,
 }
