+++ l10n_fr_department_delivery/readme/DESCRIPTION.rst
@@ -1,0 +1,3 @@
+This module improves the availability of a shipping method based on *french departments*.
+
+It adds a field *French Departments* on the *Destination Availability* tab of *Shipping Methods*.
